move as pure cut/paste
