+++ statsmodels/tools/linalg.py
@@ -10,6 +10,7 @@ from statsmodels.compat.python import lmap, range
 import numpy as np
 from scipy.linalg import svd as decomp_svd
 from scipy.linalg.lapack import get_lapack_funcs
+from scipy import sparse
 from numpy import asarray, zeros, sum, conjugate, dot, transpose
 import numpy
 from numpy import asarray_chkfinite, single
@@ -267,6 +268,167 @@ def stationary_solve(r, b):
 
     return x
 
+
+def _dot(x, y):
+    """
+    Returns the dot product of the arrays, works for sparse and dense.
+    """
+
+    if isinstance(x, np.ndarray) and isinstance(y, np.ndarray):
+        return np.dot(x, y)
+    elif sparse.issparse(x):
+        return x.dot(y)
+    elif sparse.issparse(y):
+        return y.T.dot(x.T).T
+
+
+# From numpy, adapted to work with sparse and dense arrays.
+def _multi_dot_three(A, B, C):
+    """
+    Find best ordering for three arrays and do the multiplication.
+
+    Doing in manually instead of using dynamic programing is
+    approximately 15 times faster.
+    """
+    # cost1 = cost((AB)C)
+    cost1 = (A.shape[0] * A.shape[1] * B.shape[1] +  # (AB)
+             A.shape[0] * B.shape[1] * C.shape[1])   # (--)C
+    # cost2 = cost((AB)C)
+    cost2 = (B.shape[0] * B.shape[1] * C.shape[1] +  # (BC)
+             A.shape[0] * A.shape[1] * C.shape[1])   # A(--)
+
+    if cost1 < cost2:
+        return _dot(_dot(A, B), C)
+    else:
+        return _dot(A, _dot(B, C))
+
+
+def _dotsum(x, y):
+    """
+    Returns sum(x * y), where '*' is the pointwise product, computed
+    efficiently for dense and sparse matrices.
+    """
+
+    if sparse.issparse(x):
+        return x.multiply(y).sum()
+    else:
+        # This way usually avoids allocating a temporary.
+        return np.dot(x.ravel(), y.ravel())
+
+
+def _smw_solver(s, A, AtA, Qi, di):
+    """
+    Returns a solver for the linear system:
+
+    .. math::
+
+        (sI + ABA^\prime) y = x
+
+    The returned function f satisfies f(x) = y as defined above.
+
+    B and its inverse matrix are block diagonal.  The upper left block
+    of :math:`B^{-1}` is Qi and its lower right block is diag(di).
+
+    Parameters
+    ----------
+    s : scalar
+        See above for usage
+    A : ndarray
+        p x q matrix, in general q << p, may be sparse.
+    AtA : square ndarray
+        :math:`A^\prime  A`, a q x q matrix.
+    Qi : square symmetric ndarray
+        The matrix `B` is q x q, where q = r + d.  `B` consists of a r
+        x r diagonal block whose inverse is `Qi`, and a d x d diagonal
+        block, whose inverse is diag(di).
+    di : 1d array-like
+        See documentation for Qi.
+
+    Returns
+    -------
+    A function for solving a linear system, as documented above.
+
+    Notes
+    -----
+    Uses Sherman-Morrison-Woodbury identity:
+        https://en.wikipedia.org/wiki/Woodbury_matrix_identity
+    """
+
+    # Use SMW identity
+    qmat = AtA / s
+    if sparse.issparse(qmat):
+        qmat = qmat.todense()
+    m = Qi.shape[0]
+    qmat[0:m, 0:m] += Qi
+    d = qmat.shape[0]
+    qmat.flat[m*(d+1)::d+1] += di
+    if sparse.issparse(A):
+        qmati = sparse.linalg.spsolve(sparse.csc_matrix(qmat), A.T)
+    else:
+        qmati = np.linalg.solve(qmat, A.T)
+
+    if sparse.issparse(A):
+        def solver(rhs):
+            ql = qmati.dot(rhs)
+            ql = A.dot(ql)
+            return rhs / s - ql / s**2
+    else:
+        def solver(rhs):
+            ql = np.dot(qmati, rhs)
+            ql = np.dot(A, ql)
+            return rhs / s - ql / s**2
+
+    return solver
+
+
+def _smw_logdet(s, A, AtA, Qi, di, B_logdet):
+    """
+    Returns the log determinant of
+
+    .. math::
+
+        sI + ABA^\prime
+
+    Uses the matrix determinant lemma to accelerate the calculation.
+    B is assumed to be positive definite, and s > 0, therefore the
+    determinant is positive.
+
+    Parameters
+    ----------
+    s : positive scalar
+        See above for usage
+    A : ndarray
+        p x q matrix, in general q << p.
+    AtA : square ndarray
+        :math:`A^\prime  A`, a q x q matrix.
+    Qi : square symmetric ndarray
+        The matrix `B` is q x q, where q = r + d.  `B` consists of a r
+        x r diagonal block whose inverse is `Qi`, and a d x d diagonal
+        block, whose inverse is diag(di).
+    di : 1d array-like
+        See documentation for Qi.
+    B_logdet : real
+        The log determinant of B
+
+    Returns
+    -------
+    The log determinant of s*I + A*B*A'.
+
+    Notes
+    -----
+    Uses the matrix determinant lemma:
+        https://en.wikipedia.org/wiki/Matrix_determinant_lemma
+    """
+
+    p = A.shape[0]
+    ld = p * np.log(s)
+    qmat = AtA / s
+    m = Qi.shape[0]
+    qmat[0:m, 0:m] += Qi
+    d = qmat.shape[0]
+    qmat.flat[m*(d+1)::d+1] += di
+    _, ld1 = np.linalg.slogdet(qmat)
+    return B_logdet + ld + ld1
 
 
 if __name__ == '__main__':
